chore: rename someone and rambo
