+++ src/event-distribution/index.ts
@@ -1,17 +1,17 @@
-import { Rambo } from "./rambo";
+import { EventDistribution } from "./event-distribution";
 import { EventHandlerOptions } from "./events";
 import { HandlerClass } from "./types/handler";
 
-const rambo = new Rambo();
+const distribution = new EventDistribution();
 
 export * from "./types/handler";
 
 export const Command = <T extends EventHandlerOptions>(options: T) => {
   return <U extends HandlerClass<T["event"]>>(target: U) => {
-    rambo.addWithOptions(options, target);
+    distribution.addWithOptions(options, target);
     return target;
   };
 };
 
-export default rambo;
+export default distribution;
 export { DiscordEvent } from "./types/base";
